Refactor `Fp` code to use `const` generics, and more (#379)

+++ ff/src/fields/utils.rs
@@ -1,6 +1,6 @@
 /// Calculates the k-adicity of n, i.e., the number of trailing 0s in a base-k
 /// representation.
-pub fn k_adicity(k: usize, mut n: usize) -> u32 {
+pub fn k_adicity(k: u64, mut n: u64) -> u32 {
     let mut r = 0;
     while n > 1 {
         if n % k == 0 {
